Add files via upload
In my recent data analysis project using Excel, I focused on examining road accidents in the UK. The project highlights various types of road accidents categorized by vehicles, offering a comprehensive view of the contributing factors and trends. I utilized a donut chart to illustrate the distribution of fatal, serious, and slight casualties, with a specific focus on casualties involving cars. A line chart was created to compare the monthly trend of casualties for the current and previous years, providing valuable insights into seasonal variations and potential areas for intervention.

Additionally, the analysis delves into casualties by different road types and surfaces, as well as by location and light conditions, offering a nuanced understanding of where and when accidents are more likely to occur. To enhance user interaction and data exploration, I included a filter panel that allows users to switch between yearly and monthly data views. Furthermore, a slicer was incorporated to differentiate between urban and rural accident data, making it easier to analyze the impact of environment on road safety. This project serves as a robust tool for understanding and mitigating road accidents in the UK.